refactor: rename nested instance module
